Manifest commands RFC
This RFC adds manifest parsing and three basic commands (sync, diff, and
status).

More error checking needs to be added. This is mostly to get some
feedback on the approach. There are some cases that turn tricky if you
always keep a local branch to avoid a detached HEAD.

I'm wondering if 'revision' is supposed to always point to a branch (as
opposed to e.g. a SHA). SHAs would be more flexible, but make it even
trickier to keep a local branch.

I've written a bit in
https://github.com/zephyrproject-rtos/zephyr/issues/6770 as well.

Signed-off-by: Ulf Magnusson <Ulf.Magnusson@nordicsemi.no>