cmd: build: fix force_cmake calculation
Checking for a forced CMake due to the presence of CMake-related
arguments should only matter if the build directory already exists and
seems to contain a zephyr build system.

Otherwise, we always need to run CMake.

Once set for any reason, the bit should be sticky, and should be a
bool.

Signed-off-by: Marti Bolivar <marti@foundries.io>