Mark top level west directory with .west_topdir file
Add an empty file .west_topdir into the top-level west directory
during initialization time, and search for it when trying to locate
that directory in the bootstrap script and west itself.

This avoids ambiguities when other directories in the installation are
also named 'west' (such as zephyr/doc/west), and propagates a fix made
in src/west/util.py to the equivalent bootstrap routine in
src/west/bootstrap/main.py (west installed in the root directory is
now found there too).

Signed-off-by: Marti Bolivar <marti@foundries.io>