Update instructions for installing a custom bootstrapper
The instructions for using an arbitrary west installation are OK, but
the docs on how to create and install a custom bootstrapper could use
some fixes. Change them to build and install the bootstrapper from a
wheel instead, and provide instructions for how to back that change
out.

(Prefer wheel here to follow the general trend treating them as the
favored package format, and also because the wheel format is
cross-platform.)

While we're here, remove the instructions on executing directly out of
the repository. You can run a custom "main" west via instructions that
are added in this patch, and it doesn't make much sense to run the
bootstrapper unless you've installed it as an entry point script
somewhere.

Signed-off-by: Marti Bolivar <marti@foundries.io>