Update instructions for installing a custom bootstrapper
The instructions for using an arbitrary west installation are OK, but
the docs on how to create and install a custom bootstrapper could use
some fixes. Change them to build and install the bootstrapper from a
wheel instead, and provide instructions for how to back that change
out.

(Prefer wheel here to follow the general trend treating them as the
favored package format, and also because the wheel format is
cross-platform.)

Signed-off-by: Marti Bolivar <marti@foundries.io>