Move bootstrapper into west namespace
Includes work by Ulf Magnusson <Ulf.Magnusson@nordicsemi.no>.

The west bootstrap code is currently being installed on the user's
system in an importable "bootstrap" module. This name is considered
likely to collide with other modules or packages in user environments,
which is undesirable.

To avoid that possibility, move the bootstrap code under the "west"
package namespace to west._bootstrap. This should not be considered to
be making the bootstrap script part of any sort of stable west API to
be used by others. It is merely a namespace pollution and general
organizational cleanup.

This move forces us to convert west to a "namespace package", which
does not include an __init__.py file. Namespace packages are defined
in PEP 420 and were added in Python 3.3.

Why?

It creates a "west._bootstrap" package under the user's site-packages,
which occurs early in sys.path. The top level west package is a
"regular package" (i.e. has an __init__.py), though. This is a
problem: regular packages must be entirely under a single directory in
the filesystem. The "main" west is not in site-packages, though: it's
fetched into the west repository cloned by "west init".

Thus, we at least must eliminate use of __init__.py in the top level
'west' package. This lets us split the subpackages and submodules
across site-packages and the init directory. For consistency, convert
the other subpackages to namespace packages as well. (This also starts
to open up opportunities for users to add their own west.runner
modules, etc.)

This conversion precludes use of relative imports, since a missing
__init__.py means there is no parent package. Relative imports are
not universally loved, anyway, so that's fine.

Finally, to finish the job, don't put any of the tests in the 'west'
package namespace, to keep "west" directories limited to the
bootstrapper/main west code. That makes for a flatter directory
hierarchy as well, which is a bonus.

Signed-off-by: Marti Bolivar <marti@foundries.io>
Signed-off-by: Ulf Magnusson <Ulf.Magnusson@nordicsemi.no>