west.util: generalize top-level directory detection
Allow finding the top level west install directory from any starting
point on the system with a new start kwarg in west_dir() and
west_topdir().

Add a new convenience predicate, in_multirepo_install(), for simply
checking whether or not a starting point is part of a multi-repo west
installation.

Signed-off-by: Marti Bolivar <marti@foundries.io>