west/main.py: only expose project commands in multi-repo installations
We need to be able to keep copying west into Zephyr's scripts/meta/
directory, because that's the only way for bug fixes and commands
which are safe for monorepo installs (like "attach") to make their way
back to users.

To make that happen, detect if the current west tree -- not the
directory from which it's called from -- is part of a multi-repo
installation, and expose the project related commands if so.

This allows a copy of west in a monorepo Zephyr to work as expected. A
multi-repo installation will be invoked either

1. by the bootstrapper
2. by a launcher shim in zephyr itself (in case zephyr-env.{sh,cmd}
   was run)

In case 1., the west install discovered by the bootstrapper is
multi-repo by construction. In case 2., the launcher shim will invoke
the multi-repo version of west in TOP/west/west, not
TOP/zephyr/scripts/meta/west, which is multi-repo as well.

Signed-off-by: Marti Bolivar <marti@foundries.io>