commands: build: fix source directory consistency check
The error message is confusing since what's output is not what was
tested. We also want to make sure to do absolute paths on both cached
and source directories just in case of issues (e.g. on Windows,
case-sensitivity is a potential source of problems).

(We can't use os.path.samefile because that's not cross-platform.)

Signed-off-by: Marti Bolivar <marti@foundries.io>