Extract manifest parsing to west.manifest
Add a new module for parsing the manifest, containing the Project
definition, as well as a routine for extracting projects from a
manifest. For now, this is a pretty straightforward copy/paste job,
but we do avoid calling log.die from the manifest parsing code by
adding a new MalformedManifest exception type.

Move the manifest schema to src/west/manifest-schema.yml as part of this.

Signed-off-by: Marti Bolivar <marti@foundries.io>