Reimplement west.manifest with more complete abstract types
Currently, west.manifest mostly can just return a list of Project
types, each of which is a namedtuple for the project contents we care
about.

It would be better and more flexible to have a "real" data type for
the manifest, along with concrete representations for its contents
such as default values and remotes.

Reimplement the module to add just these things, adding Manifest,
Defaults, and Remote types, and re-working the Project namedtuple into
a class type. All of these (except Manifest) use __slots__ to restrict
their attributes.

Keep the current users and test cases up to date, extending the tests
a bit more as well. This moves a bit more code out of project.py into
west.manifest.

Signed-off-by: Marti Bolivar <marti@foundries.io>