main.py: set ZEPHYR_BASE whenever possible
Set ZEPHYR_BASE in the environment for the duration of the command if
possible.

In order of decreasing precedence, get ZEPHYR_BASE from:

1. The west command line
2. The manifest file project named 'zephyr'
3. The calling environment's ZEPHYR_BASE

If the calling environment's ZEPHYR_BASE differs from the value
actually used, emit a warning. The user has probably run a zephyr-env
script in another Zephyr directory and forgotten about that, and will
almost certainly not like what they get.

Hopefully over time we can deprecate the zephyr-env scripts entirely,
but that remains to be seen.

Fixes: #46

Signed-off-by: Marti Bolivar <marti@foundries.io>