west.manifest: forbid duplicate project paths
While implementing a manifest-based scan for the zephyr project, I
noticed that we aren't ensuring that projects always have distinct
paths (it's generally OK for them to have non-overlapping names; they
just can't be cloned to the same directories on the file system).

Add a check for that in west.manifest, along with a regression test.

Signed-off-by: Marti Bolivar <marti@foundries.io>