commands: project: Check if projects are up-to-date when rebasing
_update() updates the special west/ and manifest/ repositories and has
code for checking if HEAD contains all the commits from a particular
branch (is up-to-date re. the branch). Move that code into a separate
_up_to_date_with() function, and also use it when rebasing normal
projects.

This gives a less spammy message when running 'pull' and 'rebase' on
already up-to-date projects:

  $ west rebase
  === zephyr (zephyr/) is up-to-date with manifest-rev

A function for checking how many commits a commit is behind/ahead
another commit would be nice to have in the future as well.

This commit also removes the special error message when self-updating
fails, returning to plain _git() with check=True (which points out the
repository path at least). I think it might be nicer to handle "context
hints" by raising an exception from _git_helper() on errors and catching
it, but I thought I'd add it separately.

Signed-off-by: Ulf Magnusson <ulfalizer@gmail.com>