Make configuration values always available
This is nice in case something tries to read a configuration value
before the configuration files have been read. This could happen for the
logging functions for example, and for tests.

The configuration is empty before the configuration files have been
read, so defaults must be provided. They must be anyway though, for the
code to be robust in case configuration files are missing.

Signed-off-by: Ulf Magnusson <Ulf.Magnusson@nordicsemi.no>