commands: project: Only update west/manifest via fast-forward
If either the west or the manifest repository has local modifications,
skip automatically updating the modified repository, with a warning.
This avoids merge conflicts when working on them.

Maybe the manifest repository update could be skipped if the west
repository is out of date, and vice versa, but I suspect this version
might be good enough.

Fixes: #59
Fixes: #62

Signed-off-by: Ulf Magnusson <Ulf.Magnusson@nordicsemi.no>