commands: project: Clarify that 'west checkout' deals with local branches
As long as that's the case, it should be documented.

I suspect it might be a feature. If you've created a branch do to some
work with 'west branch fix-stuff repo1 repo2 repo3', then it's probably
unexpected if 'west checkout fix-stuff' checks out a remote branch
'fix-stuff' that happens to exist in some other repository.

Fixes: #34

Signed-off-by: Ulf Magnusson <Ulf.Magnusson@nordicsemi.no>