commands: list: accept project arguments
Support a list of projects in the arguments. The listed projects will
be limited to those named, in the order they are given on the command
line. This is the same way that "fetch", "forall", and other commands
operate on all projects by default, or just specified projects.

Remove the special-case text printing the manifest path. We'll handle
that by treating west and the manifest as projects in a later patch.

Signed-off-by: Marti Bolivar <marti@foundries.io>