commands: list: add --format, --all options
Add a -f / --format option that allows users to list projects exactly how
they want them.

Also add a -a / --all flag that does not ignore the west or manifest
repositories in the listing. This is by analogy with the ls command;
we are treating west and the manifest as if they were hidden.

Signed-off-by: Marti Bolivar <marti@foundries.io>