cmake: west: Take the 'kernel' name out of the .bin/.hex/.elf file
Rename variables to make west more agnostic, and therefore potentially
more flexible, to what it is flashing.

This is a small step towards solving #7868 and other similair issues
in Zephyr. Essentially, sometimes, for instance when a bootloader is
used, or a dual-core platform is used, we need to flash multiple
images/application's, not just the 'kernel'.

To prepare 'west' for this future feature we make it agnostic to what
kind of hex/bin/elf file it is flashing by taking the 'kernel' name
out of the variables. This allows the hex file to contain not just the
kernel, but also a bootloader, or a second image.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>