tests: project: Add tests for 'west update'
Create some dummy manifest, west, and project upstream repositories, and
a simple configuration file that points to them. Use this setup to test
--reset/update-manifest, --reset/update-west, and --reset-projects.

The --reset-* flags are used by 'west init' when reinitializing an
existing west installation. Also test that the correct --reset-* flags
are generated when calling the bootstrapper to reinitialize an
installation.

Some simple helper functions were added for creating repositories,
adding commits, and getting the subject of a repository's HEAD commit.

Signed-off-by: Ulf Magnusson <Ulf.Magnusson@nordicsemi.no>