Be more helpful to newcomers.
The error output when the bootstrapper is called without arguments or
with -h is lacking. Try to improve it.

Tweak the output when called without arguments when an installation is
found as well, in the same spirit.

Finally, change the print_usage() to a print_help() when no command is
given, to get additional context.

Fixes: #108

Signed-off-by: Marti Bolivar <marti@foundries.io>