runners: jlink: debug log commander script contents
We intend that 'west -v flash' provides sufficient information for the
motivaed user to stop using west if they so choose. The use of a
temporary file in the jlink runner whose contents are deleted as west
exits prevents that. Print its contents so we can let people get back
to work without west.

Fixes: #113

Signed-off-by: Marti Bolivar <marti@foundries.io>