manifest: Support both west and manifest sections
Rework the manifest parsing in order to support both the manifest and
west sections that can now be present in the yaml file.
This enables the manifest to be parsed fully and include both the
project information from the manifest section and also the special
metadata associated with west itself.

Signed-off-by: Carles Cufi <carles.cufi@nordicsemi.no>