bootstrap: don't check that src is a URL in clone()
Git clone can take a raw file system path instead of a URL when
cloning a repository that is already available locally somewhere else.

When this happens, git tries to use hard links to create the files in
the destination, which is more efficient, so the limitation to URLs is
not good. Remove it.

Signed-off-by: Marti Bolivar <marti@foundries.io>