commands: group related commands together
Some utilities which use subcommands break them into groups in their
help output, to make it easier for the reader to find what they are
looking for. Two notable examples are docker and git.

Do the same thing for west, by tweaking the COMMANDS type to group
related commands together under a string key which describes the
group, and adding a custom argument parser subclass which uses this
information when formatting the top-level command help.

Add a group of "hidden" commands which is just PostInit for
now. Eventually that will be going away, and in the meantime, we'd
like to sweep it under the rug.

I would have preferred to do this by hooking into argparse itself, but
it is annoyingly opaque about what constitutes stable API.

The top-level "west -h" output now looks like this (without ">"s):

> $ west -h
> usage: west [-h] [-z ZEPHYR_BASE] [-v] [-V] <command> ...
>
> The Zephyr RTOS meta-tool.
>
> optional arguments:
>   -h, --help            show this help message and exit
>   -z ZEPHYR_BASE, --zephyr-base ZEPHYR_BASE
>                         Override the Zephyr base directory. The
>                         default is the manifest project with path
>                         "zephyr".
>   -v, --verbose         Display verbose output. May be given multiple
>                         times to increase verbosity.
>   -V, --version         print the program version and exit
>
> west commands used in various situations:
>  building and running zephyr:
>      build:             compile a Zephyr application
>      flash:             flash and run a binary on a board
>      debug:             flash and interactively debug a Zephyr
>                         application
>      debugserver:       connect to board and launch a debug server
>      attach:            interactively debug a board
>
>  managing multiple repositories in the installation:
>      list:              print information about projects in the west
>                         manifest
>      clone:             clone remote projects into local installation
>      fetch:             "git fetch" changes from project remotes
>      pull:              "git pull" changes from project remotes
>      rebase:            "git rebase" local projects onto manifest
>                         versions
>      branch:            create a branch in one or more local projects
>      checkout:          check out a branch in one or more local
>                         projects
>      diff:              "git diff" for one or more projects
>      status:            "git status" for one or more projects
>      update:            update the manifest and west repositories
>      forall:            run a command in one or more local projects
>
> Run "west <command> -h" for help on each command.

The command-level help (e.g. "west build -h") is not changed.

Fixes: #107
Signed-off-by: Marti Bolivar <marti@foundries.io>